requirements.txt
aio-pika==9.3.0
aiormq==6.7.7
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
databases==0.8.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
pamqp==3.2.1
pika==1.3.2
psycopg2==2.9.9
pydantic==2.5.1
pydantic_core==2.14.3
pyproject==1.3.1
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.24.0.post1
yarl==1.9.2
